requirements.txt
allure-pytest==2.11.1
allure-python-commons==2.11.1
asgiref==3.5.2
async-generator==1.10
atomicwrites==1.4.1
attrs==22.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Django==4.1.2
django-bootstrap4==22.2
django-pytest==0.2.0
exceptiongroup==1.0.0
execnet==1.9.0
Flask==2.2.2
h11==0.14.0
idna==3.4
importlib-metadata==5.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
outcome==1.2.0
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
pytest==6.2.5
pytest-base-url==2.0.0
pytest-forked==1.4.0
pytest-html==3.2.0
pytest-metadata==2.0.3
pytest-selenium==4.0.0
pytest-variables==2.0.0
pytest-xdist==2.5.0
pytz==2022.5
requests==2.28.1
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tenacity==6.3.1
toml==0.10.2
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.5
urllib3==1.26.12
Werkzeug==2.2.2
wsproto==1.2.0
zipp==3.9.0